pyproject.toml
[tool.poetry]
name = "akee-predict"
version = "0.1.0"
description = ""
authors = ["Lewis Tham <lewistham9x@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
apache-airflow = "^2.5.3"
pandas = "^1.5.3"
google-cloud-bigquery = "2.28.0"
python-dotenv = "^0.21.1"
google-cloud-storage = "^2.7.0"
snscrape = "^0.5.0.20230113"
apache-airflow-providers-apache-beam = "^4.2.0"
cloudscraper = "^1.2.68"
apache-airflow-providers-google = "^9.0.0"
xgboost = "^1.7.5"
lightgbm = "^3.3.5"
imblearn = "^0.0"
joblib = "^1.2.0"
pip = "^23.1"
install = "^1.3.5"
scikit-learn = "^1.2.2"
jupyter = "^1.0.0"
ipykernel = "^6.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
